Break out DefaultRichLink into it's own file
