+++ src/web/lib/ElementRenderer.tsx
@@ -5,7 +5,7 @@ import { CalloutBlockComponent } from '@root/src/web/components/elements/Callout
 import { CaptionBlockComponent } from '@root/src/web/components/elements/CaptionBlockComponent';
 import { CommentBlockComponent } from '@root/src/web/components/elements/CommentBlockComponent';
 import { CodeBlockComponent } from '@root/src/web/components/elements/CodeBlockComponent';
-import { DefaultRichLink } from '@root/src/web/components/RichLink';
+import { DefaultRichLink } from '@root/src/web/components/DefaultRichLink';
 import { DocumentBlockComponent } from '@root/src/web/components/elements/DocumentBlockComponent';
 import { DisclaimerBlockComponent } from '@root/src/web/components/elements/DisclaimerBlockComponent';
 import { DividerBlockComponent } from '@root/src/web/components/elements/DividerBlockComponent';
